+++ tests/integration/x_erc20_test.go
@@ -1,0 +1,13 @@
+package integration
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/suite"
+
+	"github.com/cosmos/evm/tests/integration/x/erc20"
+)
+
+func TestERC20GenesisTestSuite(t *testing.T) {
+	suite.Run(t, erc20.NewGenesisTestSuite(CreateKiichain))
+}